refactor file list into component
