+++ src-tauri/.gitignore
